assignment-2/data-format-translation/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["DryDish <57399961+DryDish@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
xmltodict = "^0.13.0"
requests = "^2.28.1"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
